pipeline_periodic/requirements.txt
pandas
PyMySQL
python-dotenv
awswrangler
boto3